pojo files

+++ src/main/java/com/product/catalog/ProductCatalog/JsonMap/Products.java
@@ -1,21 +1,22 @@
 package com.product.catalog.ProductCatalog.JsonMap;
 
-import java.util.List;
 import com.fasterxml.jackson.annotation.JsonProperty;
+
+import java.util.List;
 
 public class Products{
 
 	@JsonProperty("shopCode")
-	private Integer shopCode;
+	private Long shopCode;
 
 	@JsonProperty("products")
 	private List<ProductsItem> products;
 
-	public void setShopCode(Integer shopCode){
+	public void setShopCode(Long shopCode){
 		this.shopCode = shopCode;
 	}
 
-	public Integer getShopCode(){
+	public Long getShopCode(){
 		return shopCode;
 	}
